refactor: Created a new file for sanity config
