+++ pages/index.tsx
@@ -6,7 +6,6 @@ import Header from '../components/Header'
 import PageHead from '../components/PageHead'
 import ProjectDisplaySection from '../components/ProjectDisplaySection'
 import SkillDisplaySection from '../components/SkillDisplaySection'
-import { createClient } from "next-sanity"
 import query from "../queries.json"
 import Link from 'next/link'
 import Background from '../components/Background'
@@ -14,11 +13,13 @@ import { homePageType } from '../types'
 import { useCallback, useEffect, useState } from 'react'
 import gsap from 'gsap'
 import ScrollTrigger from 'gsap/dist/ScrollTrigger'
+import { sanityClient } from "../utils/sanity_config"
 
 
 const Home: NextPage<homePageType> = (
   { landing_section, about_data, social_medias, skill_list, projects }
 ) => {
+  // page animation
   const [timeline,] = useState(() => gsap.timeline({
     defaults: {
       ease: "slow(0.3, 0.4, false)",
@@ -99,13 +100,6 @@ const Home: NextPage<homePageType> = (
   )
 }
 
-
-export const sanityClient = createClient({
-  projectId: process.env['NEXT_PUBLIC_SANITY_PROJECT_KEY'] ?? "",
-  dataset: "production",
-  useCdn: process.env.NODE_ENV === "production",
-  apiVersion: "v2021-08-31",
-})
 
 export async function getStaticProps() {
