+++ pages/project/[id].tsx
@@ -2,8 +2,7 @@ import gsap from 'gsap'
 import ScrollTrigger from 'gsap/dist/ScrollTrigger'
 import { GetStaticPropsContext } from 'next'
 import Link from 'next/link'
-import { useCallback, useEffect, useState } from 'react'
-import { sanityClient } from '..'
+import { useEffect } from 'react'
 import Background from '../../components/Background'
 import Carousel from '../../components/Carousel'
 import FeatureContent, { FeatureContentImage } from '../../components/FeatureContent'
@@ -11,6 +10,7 @@ import Header from '../../components/Header'
 import OtherProjectDisplay from '../../components/OtherProjectDisplay'
 import PageHead from '../../components/PageHead'
 import TechStackDisplay from '../../components/TechStackDisplay'
+import { sanityClient } from '../../utils/sanity_config'
 
 import { projectPageType } from '../../types'
